Exposed `skyFormatDate` function
