+++ src/app/public/index.ts
@@ -8,3 +8,5 @@ export * from './modules/date-range-picker/types/date-range-calculator-id';
 export * from './modules/date-range-picker/types/date-range-calculator-type';
 export * from './modules/date-range-picker/date-range-picker.module';
 export * from './modules/date-range-picker/date-range.service';
+
+export * from './modules/shared/format-date';
